requirements/common.txt
Django==3.1.5
djangorestframework==3.12.2
django-environ==0.4.5
requests==2.25.1
python-dateutil==2.8.1
isodate==0.6.0
requests==2.25.1
pytest-cov==2.11.1
redis==3.5.3
djangorestframework-dataclasses==0.8
python-json-logger==2.0.1
django-cors-headers==3.7.0